Complete logic for FindClientCommand and update its javadoc.
